Bugfix: Avoid integer overflow in o5m parser
See #329

+++ include/osmium/io/detail/o5m_input_format.hpp
@@ -166,7 +166,7 @@ namespace osmium {
                 }
 
                 bool ensure_bytes_available(std::size_t need_bytes) {
-                    if ((m_end - m_data) >= static_cast<int64_t>(need_bytes)) {
+                    if (static_cast<std::size_t>(m_end - m_data) >= need_bytes) {
                         return true;
                     }
